requirements.txt
snakefood==1.4
click==6.7
pygithub
pip
gitpython
python-json-logger